chore(TimerTrigger): Move code from `main.rs` to `lib.rs` With moving everything but the main function to  we make the  importable for other projects
