+++ examples/spin-timer/trigger-timer.json
@@ -1,9 +1,9 @@
 {
     "name": "trigger-timer",
-    "description": "IT IS I THE GREAT OZ",
-    "homepage": "www.example.com",
+    "description": "Run Spin components at timed intervals",
+    "homepage": "https://github.com/fermyon/spin/tree/main/examples/spin-timer",
     "version": "0.1.0",
-    "spinCompatibility": ">=0.7",
+    "spinCompatibility": ">=2.0",
     "license": "Apache-2.0",
     "packages": [
         {
